+++ src/config.template.json
@@ -1,0 +1,4 @@
+{
+  "gitlabApi": "https://gitlab.example.com/api/v4",
+  "privateToken": "ABCDEF123456in default"
+}
